+++ configs.py
@@ -74,7 +74,10 @@ PATH    = {
 
 
     ## Paths associated with Paras Selection
-    'processed_data'                : "./backup/processed_data/[SPLIT]/data_[N_SHARD].pkl",
+    'processed_data'                : "./backup/processed_data/[SPLIT]/data_[N_SHARD].csv",
     'data_training'                 : "./backup/data_parasselection/[SPLIT]/data_[N_SHARD].csv",
     'savemodel_ParasSelection'      : "./backup/saved_model/paras_selector.pt"
+
+    ## Paths after ParasSelection
+    'data'                          : "./backup/data_[SPLIT].csv"
 }